requirements.txt
flask
flask-pymongo
python-dotenv
reportlab
fpdf2
openai